Move all views to com.twilio.chat.demo.views
