+++ build.gradle
@@ -14,5 +14,8 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
+        maven {
+            url  "https://oss.sonatype.org/content/repositories/snapshots"
+        }
     }
 }
