+++ chat-demo-android/build.gradle
@@ -1,19 +1,20 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+apply plugin: 'kotlin-android-extensions'
 
 def ACCESS_TOKEN_SERVICE_URL = project.hasProperty('ACCESS_TOKEN_SERVICE_URL') ? "\""+project.getProperty('ACCESS_TOKEN_SERVICE_URL')+"\"" : "\"http://example.com/token\""
 
 dependencies {
-    compile 'com.android.support:support-v4:23.3.0+'
+    compile 'com.android.support:support-v4:25.1.0+'
     compile 'com.google.firebase:firebase-messaging:10.2.0' // FCM
     compile 'com.google.android.gms:play-services-base:10.2.0'
-    compile 'uk.co.ribot:easyadapter:1.5.0@aar'
     compile 'com.jakewharton.timber:timber:4.5.1'
-    compile 'com.jakewharton:butterknife:8.6.0'
+    compile 'com.jakewharton:kotterknife:0.1.0-SNAPSHOT'
+    compile 'eu.inloop:simplerecycleradapter:0.3.4'
     compile 'com.twilio:chat-android:2.0.0'
     compile 'com.twilio:accessmanager-android:0.1.0'
     compile "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.6.0'
 }
 
 android {
