+++ chat-demo-android/src/main/java/com/twilio/chat/demo/views/ChannelViewHolder.kt
@@ -1,4 +1,4 @@
-package com.twilio.chat.demo
+package com.twilio.chat.demo.views
 
 import com.twilio.chat.demo.ChannelModel
 import com.twilio.chat.Channel.ChannelStatus
@@ -9,22 +9,23 @@ import android.graphics.Color
 import android.util.Log
 import android.view.View
 import android.widget.TextView
-import butterknife.BindView
-import butterknife.ButterKnife
+import com.twilio.chat.demo.R
 import uk.co.ribot.easyadapter.ItemViewHolder
 import uk.co.ribot.easyadapter.PositionInfo
 import uk.co.ribot.easyadapter.annotations.LayoutId
 import uk.co.ribot.easyadapter.annotations.ViewId
+import kotlinx.android.synthetic.main.channel_item_layout.*
 
 @LayoutId(R.layout.channel_item_layout)
 class ChannelViewHolder(internal var view: View) : ItemViewHolder<ChannelModel>(view) {
-    internal @BindView(R.id.channel_friendly_name) lateinit var friendlyName: TextView
-    internal @BindView(R.id.channel_sid) lateinit var channelSid: TextView
-    internal @BindView(R.id.channel_updated_date) lateinit var updatedDate: TextView
-    internal @BindView(R.id.channel_created_date) lateinit var createdDate: TextView
-    internal @BindView(R.id.channel_users_count) lateinit var usersCount: TextView
-    internal @BindView(R.id.channel_total_messages_count) lateinit var totalMessagesCount: TextView
-    internal @BindView(R.id.channel_unconsumed_messages_count) lateinit var unconsumedMessagesCount: TextView
+    val friendlyName: TextView by bindView(R.id.channel_friendly_name)
+//    internal @BindView(R.id.channel_friendly_name) lateinit var
+//    internal @BindView(R.id.channel_sid) lateinit var channelSid: TextView
+//    internal @BindView(R.id.channel_updated_date) lateinit var updatedDate: TextView
+//    internal @BindView(R.id.channel_created_date) lateinit var createdDate: TextView
+//    internal @BindView(R.id.channel_users_count) lateinit var usersCount: TextView
+//    internal @BindView(R.id.channel_total_messages_count) lateinit var totalMessagesCount: TextView
+//    internal @BindView(R.id.channel_unconsumed_messages_count) lateinit var unconsumedMessagesCount: TextView
 
     override fun onSetListeners() {
         view.setOnClickListener {
@@ -34,8 +35,6 @@ class ChannelViewHolder(internal var view: View) : ItemViewHolder<ChannelModel>(
     }
 
     override fun onSetValues(channel: ChannelModel, arg1: PositionInfo) {
-        ButterKnife.bind(view)
-
         friendlyName.text = channel.friendlyName
         channelSid.text = channel.sid
